this is the first commit of the branch. I have a dummy app with the login, register, and home views
