+++ RatPack_App/app/src/main/java/edu/ratpack/nikitarajput/cs2340/gatech/ratpack_app/Login_Activity.java
@@ -1,0 +1,13 @@
+package edu.ratpack.nikitarajput.cs2340.gatech.ratpack_app;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class Login_Activity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login_);
+    }
+}
